Update sub_modules.py to remove UserWarning
The nbeats/sub_module.py raises the following warning:
UserWarning: Creating a tensor from a list of numpy.ndarrays is extremely slow. Please consider converting the list to a single numpy.ndarray with numpy.array() before converting to a tensor.
This commit replaces the relevant sections with the appropriate code to remove the warning. (similar to #754)