[MNT] CI runs without soft dependencies (#1638)
This adds test runs without soft dependencies.

The run covers currently removal of the two extant soft dependency sets, `graph` and `mqf2`. This tests isolation in the extents of test coverage, and also isolation of the tests themselves requiring these.

The setup can be extended in the future to further soft dependency sets, testing that the code functions in their absence.