[ENH] refactor `__init__` modules to no longer contain classes (#1738)
This PR refactors the `__init__` modules in `models`, moving any content
such as neural network classes to submodules.

Imports will continue working unchanged, as the native contents of the
module are still exported.

This is preparing a further structural refactors.

The moved files are also not changed, in order to retain the commit as
one that moves files, hence retaining edit history. In order for this to
take effect on `main`, PR
https://github.com/sktime/pytorch-forecasting/pull/1739 must be merged
first (even though this will result in an intermediate, non-passing
state).