[MNT] update nbeats/sub_modules.py to remove overhead in tensor creation (#1580)
This commit  removes a warning raised by nbeats/sub_module.py: 
`UserWarning: Creating a tensor from a list of numpy.ndarrays is
extremely slow. Please consider converting the list to a single
numpy.ndarray with numpy.array() before converting to a tensor.`
This commit replaces the relevant sections with the appropriate code to
remove the warning in the nbeats submodule. (similar to
https://github.com/jdb78/pytorch-forecasting/issues/754)

Co-authored-by: Felix Hirwa Nshuti <hirwanshutiflx@gmail.com>