[MNT] Temporary fix for lint errors to conform to the recent changes in linting rules see #1749 (#1748)
### Description

This PR fixes the linting errors after the lint rules were updated
recently.

Initially added a lot of `noqa` to have a baseline that passes the lint
check going to work on removing all `noqa` to the minimal value. This is
still a work in progress.

This temporary fix is for allowing PRs to be in mergeable states given
that they pass code quality checks.
The exhaustive refactoring is being done in #1749 to ensure the code
quality.