[MNT] move model base classes into `models.base` module - part 1 (#1773)
This PR moves the model base classes into the `modeles.base` module,
which is a folder module with private submodules.

This is a pure refactor, preparing for a `base` module with more files -
for instance, towards
https://github.com/sktime/pytorch-forecasting/issues/1771.

Part 1 of 2, part 2 in
https://github.com/sktime/pytorch-forecasting/pull/1774, separate in
order to ensure that files are properly tracked as renamed.