[ENH] two-way linkage of model package classes and neural network classes (#1888)
Stacks on https://github.com/sktime/pytorch-forecasting/pull/1892

This PR adds a `pkg` property to models that is also tested to be
inverse to `get_model_cls` interface of the package class.

To be discussed if this is a good idea - as previously discussed in the
meet-up, the exact use case for the `pkg` method is unclear.