[ENH] Add a copy of `BaseFixtureGenerator` to `pytorch-forecasting/tests/_base` as a true base class (#1919)
#### What does this implement/fix? Explain your changes.

Currently we use the `BaseFixtureGenerator` for generating fixtures for
tests involving estimators and recently metrics (#1907). But the current
import location is inside `test_all_estimators.py`. This PR aims to make
this class a true "base" class by moving it to a new location. This will
be stacked upon by #1907 and #1908 to respectively implement child
classes which inherit from this base class.

The PR serves simply as the base for serving looped, parametrized tests for metrics and models.