[ENH] Improve test framework for v1 metrics (#1907)
#### Reference Issues/PRs

Fixes #1904.

#### What does this implement/fix? Explain your changes.

Implements a skbase fixture generator class for common tests for all
metrics in v1. Common testing is successful on all metrics except 1 -
`MASE`(point). These special metrics are addressed in their own tests
under `test_metrics.py`.

This PR also makes changes to the source location of SkbaseBaseObject -
`_BaseObject` and is now shared by `test_all_estimators` and
`test_all_metrics`.

This PR has 4 aspects:

- `BasePtMetric` : an `_SkbaseBaseObject` class that allows the metric
package containers to be discovered for testing. It implements the class
methods for retrieving the name, the `class` of the metric and the set
of test_params for testing a metric. It also implement the method which
retrieves the exact name of the data generation pytest fixture for each
metric.

- Individual modules for the package containers of each metric. Derives
from `BasePtMetric` and contains overriden methods from the base class.

- `TestAllPtMetrics` is the main testing class of the library. it
implements the integration test for metrics and test for reduction of
computed loss.

- `MetricFixtureGenerator` is a child class of `BaseFixtureGenerator`
for generating fixtures for the tests. It implement a separate case of
`_generate_object_instance` to allow parametrization of the test_params
passed into a metric during its initialisation

Exception and skipped tests due to non-conformance:

1) MQF2DistributionLoss - skipped due to non-conformance with the design
for `to_prediction`, requires us to pass its own prediction_length and
not the `max_prediction_length` from the model.
2) PoissonLoss - it's implementation of `to_quantiles` is non-conformant
with the expectation of point metric. It returns a 3d quantile dimension
(equal to the len(quantiles))` when a 2d tensor is passed into a call of
`to_quantiles` with `PoissonLoss`. For point metrics this is usually
`(batch_size, prediction_length, 1)`.
3) MultivariateNormalDistribution - This metric is compliant with most
of the existing contracts for the API. It is non-compliant only with the
`loss()` method API, where the `loss()` returns a scalar value unlike
all the remaining metrics which return either a 2d tensor or a 3d tensor
(distribution - 2d tensor with log prob calculated for batch and
timesteps