[BUG] Device inconstency in `MQF2DistributionLoss` raising: RuntimeError: Expected all tensors to be on the same device (#1916)
Fixes #1182

In the current implementation, the `picnn` is initialized during class
construction, and the device that it defaults to isn't being updated
when the model is moved to another device.

* Added device movement method `to()` to ensure that `picnn` is moved
along with the loss function.
* Added automatic device sync in `map_x_to_distribution` to ensure that
`picnn` is on the same device as the input tensor.

Also added the tests mocking the accelerators on a high level to test
the synchronization of devices within this class.