[ENH] Add missing test for forward output of `TimeXer` as proposed in #1936 (#1951)
#### What does this implement/fix? Explain your changes.

Adds the test for the output shapes of the forward method of `TimeXer`.
The test checks the output contract for the model, which is being
implemented in #1936. Currently skipped under `test_timexer.py` due to
obvious causes of failure i.e current version of model on main is still
not updated to the required output contract.